requirements.txt
# Flask Framework
Flask==3.0.0
Flask-CORS==4.0.0

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9  # PostgreSQL (optionnel)

# Security
bcrypt==4.1.2
PyJWT==2.8.0

# Utilities
python-dotenv==1.0.0

# IA 
groq 

# Analyse document
PyPDF2

# Development (optionnel)
pytest==7.4.4
pytest-flask==1.3.0